requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pyjwt==2.8.0
scikit-learn==1.3.2
numpy==1.24.3